% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkoping_pop_api.R
\name{linkoping_pop}
\alias{linkoping_pop}
\title{API for the population in Linkoping population from year 1970 and 2020}
\usage{
linkoping_pop()
}
\value{
Returns a data.frame with population data in Linköping
}
\description{
The function returns a data.frame with variables year, gender and the population count.
}
\examples{
pop <- linkoping_pop()
}
